pyproject.toml
[tool.poetry]
name = "dz-27"
version = "0.1.0"
description = ""
authors = ["vladimirmac <vladimirbodrin89@gmail.com>"]
readme = "README.md"
packages = [{include = "dz_27"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
psycopg2-binary = "^2.9.5"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
python-dateutil = "^2.8.2"
rich = "^13.3.1"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
